Add a git module under librad-test
We often need to make git operations such as committing to a working
copy. This is the inception of that.
